07-AI-Bot/requirements.txt
livekit-agents~=0.8
livekit-plugins-deepgram~=0.6
livekit-plugins-silero~=0.6
transformers
pyttsx3
torch